+++ SVRichPresence/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Once again removed Mac and Linux auto-setup.
+
+### Fixed
+- Fixed crashing from Discord Game SDK by changing libraries.
 
 ## [2.3.0] - 2019-12-05
 ### Added